Code scanning fixes (#124)
* Unless checked, the resulting number can exceed the expected range (CWE-190)

* getchar - Check buffer boundaries if used in a loop including recursive loops (CWE-120, CWE-20)

* sprintf does not check for buffer overflows (CWE-120)

* Always make use of the configured logging level

* strcat does not check for buffer overflows

* strcpy does not check for buffer overflows

* buffer - statically-sized arrays can be improperly restricted

* memcpy checks

* Fix uninitialized variable

* Fix using pointer to local variable that is out of scope

* Avoid syntax errors on static analysers due to macros

* Address several cppcheck warnings

* Fix shadows outer variable

* Fix shadows outer argument

* Minor argument name change

* Properly format string for ulong

* Use pointers to const when referring to string literals

* Do not use the rand() function for generating pseudorandom number

* Do not use atoi()

* Use setvbuf instead of setbuf

* Do not cast away const

* Explicitly set the null-terminator after strncpy

* MISRA 3.1 rule

* MISRA 10.1 rule

* MISRA 10.4 rule

* MISRA 10.8 rule

* MISRA 11.5 rule

* MISRA 12.1 rule

* MISRA 12.3 rule

* MISRA 13.3 rule

* MISRA 13.4 rule

* MISRA 14.2 rule

* MISRA 14.4 rule

* MISRA 15.1 rule

* MISRA 15.2 rule

* MISRA 15.5 rule

* MISRA 15.6 rule

* MISRA 15.7 rule

* MISRA 16.3 rule

* MISRA 16.4 rule

* MISRA 16.6 rule

* MISRA 17.7 rule

* MISRA 17.8 rule

* MISRA 18.8 rule

* MISRA 20.1 rule

* MISRA 21.10 rule

* MISRA 21.6 rule

* MISRA 21.8 rule

* Force GitHub actions to test serial and bluetooth builds on supported ports

+++ src/utils/checksum.c
@@ -17,9 +17,9 @@
 uint32_t _z_crc32(const uint8_t *message, size_t len) {
     uint32_t crc = 0xFFFFFFFF;
     for (size_t i = 0; i < len; i++) {
-        crc = crc ^ message[i];
+        crc = crc ^ (uint32_t)message[i];
         for (int j = 0; j < 8; j++) {
-            crc = (crc >> 1) ^ (0x04C11DB7 & (-(crc & 1)));
+            crc = (crc >> 1) ^ ((uint32_t)0x04C11DB7 & (-(crc & (uint32_t)1)));
         }
     }
     return ~crc;
